Remove custom PATH setting
This commit removes the explicit PATH setting in favor of the system's
default $PATH. The explicit setting is a remnant from getting setup
help at Puppet and I'm not sure it's necessary. The path elements are
already part of the default $PATH (though maybe not exactly in the
same order), so using that is probably good enough. If future PATH
changes are necessary, they can be added later.