Remove optional rbenv configuration
Previously, rbenv would write its configuration steps to .zshrc.
As of v1.3.0, this now gets written to .zprofile by default, for use
in non-interactive shells as well as interactive ones.
If your .zshrc contains 'rbenv' anywhere, the configuration will be
added there, rather than to .zprofile, for backwards compatibility.
However, if your .zshrc contains 'rbenv init' (even in a comment, as
it was here), rbenv will conclude that the configuration has already
been added and won't make any additional changes.
Rather than leaving this in and requiring it to be uncommented,
removing it allows rbenv to perform its default configuration
process. Since we aren't customizing that in any way, there is no need
to save it for use in future configs.